+++ src/main/java/net/soulsweaponry/items/gun/GatlingGun.java
@@ -50,7 +50,12 @@ public class GatlingGun extends GunItem {
 
     @Override
     public int getBulletsNeeded(ItemStack stack) {
-        return EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0 ? super.getBulletsNeeded(stack) : (int) ConfigConstructor.gatling_gun_bullets_needed;
+        return EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0 ? this.getBulletsNeededWithInfinity(stack) : (int) ConfigConstructor.gatling_gun_bullets_needed;
+    }
+
+    @Override
+    public int getBulletsNeededWithInfinity(ItemStack stack) {
+        return (int) ConfigConstructor.gatling_gun_bullets_needed_with_infinity;
     }
 
     @Override