+++ src/main/java/net/soulsweaponry/items/gun/HunterCannon.java
@@ -49,7 +49,12 @@ public class HunterCannon extends GunItem {
 
     @Override
     public int getBulletsNeeded(ItemStack stack) {
-        return EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0 ? super.getBulletsNeeded(stack) : (int) ConfigConstructor.hunter_cannon_bullets_needed;
+        return EnchantmentHelper.getLevel(Enchantments.INFINITY, stack) > 0 ? this.getBulletsNeededWithInfinity(stack) : (int) ConfigConstructor.hunter_cannon_bullets_needed;
+    }
+
+    @Override
+    public int getBulletsNeededWithInfinity(ItemStack stack) {
+        return (int) ConfigConstructor.hunter_cannon_bullets_needed_with_infinity;
     }
 
     @Override